+++ internal/commands/loadbalancer/delete.go
@@ -1,13 +1,19 @@
 package loadbalancer
 
 import (
+	"errors"
 	"fmt"
+	"net/http"
+	"time"
 
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/completion"
+	"github.com/UpCloudLtd/upcloud-cli/v3/internal/config"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/output"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/resolver"
+	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud"
 	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud/request"
+	"github.com/spf13/pflag"
 )
 
 // DeleteCommand creates the "loadbalancer delete" command
@@ -26,12 +32,44 @@ type deleteCommand struct {
 	*commands.BaseCommand
 	resolver.CachingLoadBalancer
 	completion.LoadBalancer
+
+	wait config.OptionalBoolean
 }
 
 func (s *deleteCommand) InitCommand() {
 	// Deprecating loadbalancer in favour of load-balancer
 	// TODO: Remove this in the future
 	commands.SetSubcommandDeprecationHelp(s, []string{"loadbalancer"})
+
+	flags := &pflag.FlagSet{}
+	config.AddToggleFlag(flags, &s.wait, "wait", false, "Wait until the Kubernetes cluster has been deleted before returning.")
+	s.AddFlags(flags)
+}
+
+func Delete(exec commands.Executor, uuid string, wait bool) (output.Output, error) {
+	svc := exec.All()
+	msg := fmt.Sprintf("Deleting load balancer %v", uuid)
+	exec.PushProgressStarted(msg)
+
+	err := svc.DeleteLoadBalancer(exec.Context(), &request.DeleteLoadBalancerRequest{
+		UUID: uuid,
+	})
+	if err != nil {
+		return commands.HandleError(exec, msg, err)
+	}
+
+	if wait {
+		exec.PushProgressUpdateMessage(msg, fmt.Sprintf("Waiting for load balancer %s to be deleted", uuid))
+		err = waitUntilLoadBalancerDeleted(exec, uuid)
+		if err != nil {
+			return commands.HandleError(exec, msg, err)
+		}
+		exec.PushProgressUpdateMessage(msg, msg)
+	}
+
+	exec.PushProgressSuccess(msg)
+
+	return output.None{}, nil
 }
 
 // Execute implements commands.MultipleArgumentCommand
@@ -40,18 +78,32 @@ func (s *deleteCommand) Execute(exec commands.Executor, arg string) (output.Outp
 	// TODO: Remove this in the future
 	commands.SetSubcommandExecutionDeprecationMessage(s, []string{"loadbalancer"}, "load-balancer")
 
+	return Delete(exec, arg, s.wait.Value())
+}
+
+func waitUntilLoadBalancerDeleted(exec commands.Executor, uuid string) error {
+	ticker := time.NewTicker(5 * time.Second)
+	defer ticker.Stop()
+
+	ctx := exec.Context()
 	svc := exec.All()
-	msg := fmt.Sprintf("Deleting load balancer %v", arg)
-	exec.PushProgressStarted(msg)
 
-	err := svc.DeleteLoadBalancer(exec.Context(), &request.DeleteLoadBalancerRequest{
-		UUID: arg,
-	})
-	if err != nil {
-		return commands.HandleError(exec, fmt.Sprintf("%s: failed", msg), err)
+	for i := 0; ; i++ {
+		select {
+		case <-ticker.C:
+			_, err := svc.GetLoadBalancer(exec.Context(), &request.GetLoadBalancerRequest{
+				UUID: uuid,
+			})
+			if err != nil {
+				var ucErr *upcloud.Problem
+				if errors.As(err, &ucErr) && ucErr.Status == http.StatusNotFound {
+					return nil
+				}
+
+				return err
+			}
+		case <-ctx.Done():
+			return ctx.Err()
+		}
 	}
-
-	exec.PushProgressSuccess(msg)
-
-	return output.None{}, nil
 }